+++ tests/unit/mocks/ImageData.ts
@@ -1,0 +1,9 @@
+export class ImageData {
+  width: number;
+  height: number;
+
+  constructor(width: number, height: number) {
+    this.width = width;
+    this.height = height;
+  }
+}