+++ tests/unit/processors/grayscale/index.ts
@@ -2,10 +2,15 @@ import * as assert from 'assert';
 import { GrayscaleProcessor } from '../../../../lib/processors/grayscale';
 
 describe('GrayscaleProcessor', () => {
-  it('should return an outputFrame', () => {
+  it('should draw to an outputFrame', () => {
     const processor = new GrayscaleProcessor();
-    const testFrame = new OffscreenCanvas(1, 1);
-    const frame = processor.processFrame(testFrame);
-    assert(!!frame);
+    const input = new OffscreenCanvas(2, 2);
+    const output = document.createElement('canvas') as any;
+    processor.processFrame(input, output);
+
+    assert.strictEqual(output.drawing.inputFrame.width, 2);
+    assert.strictEqual(output.drawing.inputFrame.height, 2);
+    assert.strictEqual(output.drawing.width, 2);
+    assert.strictEqual(output.drawing.height, 2);
   });
 });